# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gamified Task Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface TaskLink {
  url: string;
  label?: string;
}

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  vpValue: number;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  calendarDate?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  estimatedMinutes?: number | null;
  /** @nullable */
  actualMinutes?: number | null;
  /** @nullable */
  links?: TaskLink[] | null;
  /** @nullable */
  notes?: string | null;
  checklistCount?: number;
  checklistCompleted?: number;
}

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  priority: TaskInputPriority;
  dueDate?: string;
  startDate?: string;
  calendarDate?: string;
  vpValue?: number;
  projectId?: number;
  /** @minimum 1 */
  estimatedMinutes?: number;
  notes?: string;
  links?: TaskLink[];
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: TaskUpdateStatus;
  priority?: TaskUpdatePriority;
  dueDate?: string;
  startDate?: string;
  calendarDate?: string;
  vpValue?: number;
  projectId?: number;
  /** @minimum 1 */
  estimatedMinutes?: number;
  /** @minimum 0 */
  actualMinutes?: number;
  notes?: string;
  links?: TaskLink[];
}

export interface TaskCompleteResult {
  task: Task;
  vpAwarded: number;
  multiplier: number;
  newTotal: number;
  tierUp: boolean;
  /** @nullable */
  newTier?: number | null;
}

export interface BulkRescheduleInput {
  /** @minItems 1 */
  taskIds: number[];
  newDate: string;
}

export interface BulkRescheduleResult {
  updated: number;
}

export interface ChecklistItem {
  id: number;
  taskId: number;
  title: string;
  completed: boolean;
  createdAt?: string;
}

export interface ChecklistItemInput {
  /** @minLength 1 */
  title: string;
}

export interface ChecklistItemUpdate {
  title?: string;
  completed?: boolean;
}

export type FocusSessionStatus = typeof FocusSessionStatus[keyof typeof FocusSessionStatus];


export const FocusSessionStatus = {
  active: 'active',
  completed: 'completed',
  abandoned: 'abandoned',
} as const;

export interface FocusSession {
  id: number;
  durationMinutes: number;
  status: FocusSessionStatus;
  /** @nullable */
  vpAwarded?: number | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export interface FocusSessionInput {
  /** @minimum 1 */
  durationMinutes: number;
}

export interface UserStats {
  totalVp: number;
  tier: number;
  tierProgress: number;
  streakDays: number;
  multiplier: number;
  tasksCompleted: number;
  focusMinutes: number;
}

export interface AnalyticsSummary {
  totalVp: number;
  tier: number;
  streakDays: number;
  tasksCompleted: number;
  focusMinutes: number;
  avgDailyVp: number;
  topPriorityCompleted: number;
}

export interface VelocityPoint {
  date: string;
  vp: number;
  tasksCompleted: number;
}

export interface Milestone {
  id: number;
  title: string;
  description: string;
  achievedAt: string;
  vpThreshold: number;
}

export interface DailyHabit {
  id: number;
  title: string;
  sortOrder: number;
  createdAt: string;
  completedToday: boolean;
}

export interface DailyHabitInput {
  title: string;
}

export interface ToggleHabitResult {
  completedToday: boolean;
}

export type ProjectType = typeof ProjectType[keyof typeof ProjectType];


export const ProjectType = {
  class: 'class',
  project: 'project',
} as const;

export interface Project {
  id: number;
  name: string;
  color: string;
  type: ProjectType;
  /** @nullable */
  emoji?: string | null;
  createdAt: string;
  taskCount?: number;
}

export type ProjectInputType = typeof ProjectInputType[keyof typeof ProjectInputType];


export const ProjectInputType = {
  class: 'class',
  project: 'project',
} as const;

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  color?: string;
  type?: ProjectInputType;
  emoji?: string;
}

export interface DashboardOverview {
  totalTasks: number;
  todoCount: number;
  inProgressCount: number;
  completedCount: number;
  criticalCount: number;
  todayTasks: Task[];
  upcomingTasks: Task[];
  userStats: UserStats;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export interface LogoutSuccess {
  success: boolean;
}

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — Bearer <sid>.
 */
export type AuthorizationSessionHeaderParameter = string;

export type ListTasksParams = {
status?: ListTasksStatus;
priority?: ListTasksPriority;
sortBy?: ListTasksSortBy;
projectId?: number;
};

export type ListTasksStatus = typeof ListTasksStatus[keyof typeof ListTasksStatus];


export const ListTasksStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export type ListTasksPriority = typeof ListTasksPriority[keyof typeof ListTasksPriority];


export const ListTasksPriority = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type ListTasksSortBy = typeof ListTasksSortBy[keyof typeof ListTasksSortBy];


export const ListTasksSortBy = {
  priority: 'priority',
  dueDate: 'dueDate',
  vpValue: 'vpValue',
  createdAt: 'createdAt',
} as const;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
};