# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gamified Task Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /focus-sessions:
    get:
      operationId: listFocusSessions
      tags: [focus]
      summary: List focus sessions
      responses:
        "200":
          description: Focus sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FocusSession"
    post:
      operationId: createFocusSession
      tags: [focus]
      summary: Start a new focus session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FocusSessionInput"
      responses:
        "201":
          description: Created focus session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FocusSession"
  /focus-sessions/{id}/complete:
    post:
      operationId: completeFocusSession
      tags: [focus]
      summary: Complete a focus session and award bonus VP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed focus session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FocusSession"
...
  /daily-habits:
    get:
      operationId: listDailyHabits
      tags: [habits]
      summary: List all daily habits with today's completion status
      responses:
        "200":
          description: List of daily habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyHabit"
    post:
      operationId: createDailyHabit
      tags: [habits]
      summary: Create a new daily habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyHabitInput"
      responses:
        "201":
          description: Created habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyHabit"
  /daily-habits/{habitId}/toggle:
    post:
      operationId: toggleDailyHabit
      tags: [habits]
      summary: Toggle today's completion for a habit
      parameters:
        - name: habitId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggle result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToggleHabitResult"
  /daily-habits/{habitId}:
    delete:
      operationId: deleteDailyHabit
      tags: [habits]
      summary: Delete a daily habit
      parameters:
        - name: habitId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects and classes
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project or class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (tasks are un-assigned)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
...
  schemas:
...
    FocusSession:
      type: object
      required: [id, durationMinutes, status, createdAt]
      properties:
        id:
          type: integer
        durationMinutes:
          type: integer
        status:
          type: string
          enum: [active, completed, abandoned]
        vpAwarded:
          type: ["integer", "null"]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    FocusSessionInput:
      type: object
      required: [durationMinutes]
      properties:
        durationMinutes:
          type: integer
          minimum: 1
...
    DailyHabit:
      type: object
      required: [id, title, sortOrder, createdAt, completedToday]
      properties:
        id:
          type: integer
        title:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
        completedToday:
          type: boolean
    DailyHabitInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
    ToggleHabitResult:
      type: object
      required: [completedToday]
      properties:
        completedToday:
          type: boolean
    Project:
      type: object
      required: [id, name, color, type, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        type:
          type: string
          enum: [class, project]
        emoji:
          type: ["string", "null"]
        createdAt:
          type: string
        taskCount:
          type: integer
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
        type:
          type: string
          enum: [class, project]
        emoji:
          type: string
...